// src/app/interfaces/User.ts
export interface IUser {
  id?: string;
  name?: string;
  lastName?: string;
  userName?: string;
  identityDocument?: number;
  age?: number;
  email?: string;
  password?: string;
  numberPhone?: number;
  country?: string;
  city?: string;
  address?: string;
  faceDescriptor?: number[]; // Arreglo de valores del descriptor facial
  faceImage?: {
   
    };
  };